escevalid: command to validate esceval output
Specifically, `escevalid` checks that each of its
arguments is one or more outputs in `esceval`
format which can be evaluated (with `eval` or
similar) predictably: in particular, not only are
they all proper single-quoted strings, but also if
a single argument would evaluate to more than one
single-quoted string, they are only separated by
one or more spaces - this protects against the
case where they are separated by a line feed,
since everything on a new line would get evaluated
as a new command by `eval` and the like (unless
the previous line ended in a backslash, which
esceval does not do).

Years ago, I realized that certain uses of esceval
could only be secure and reliable if they could
check that the output was still safe to evaluate.
I started both a C and /bin/sh implementation of
`escevalid` back then, but didn't finish until now.