esceval*.sh: remove polyfill guards
This made sense if the scripts were meant to be
pasted as polyfills, but as drop-in executables
these is-esceval*-already-available checks would
just recurse indefinitely.

In principle, I increasingly think that most code
should not polyfill things directly - it should
just rely on those things existing, and it is the
job of the system/runtime/packaging/compiler to
add the polyfills.

In my own use, I just have esceval* installed in
my PATH at all times anyway, and I've somewhat
moved on from the idea of building scripts for
which need to be complex enough to need esceval*.
It's possible it will come up, but seems unlikely.
(And I would remember to come back here to grab
the polyfill guards in any case.)

For others, even if a script does copy code from
esceval*.sh, the performance hit relative to the
C implementation is only significant for huuuuge
arguments or argument lists. So might as well
just use the in-shell implementations even if a
faster implementation is installed externally,
at least until performance becomes a problem.