escevalidenv: escevalid on environment variables
The biggest theoretical value of using environment
variables is security. Program arguments are a
global information leak on most systems, where `ps`
or /proc or other system calls can be used to see
the command line of any user's currently running
processes. Environment variables of a program can
only be read by the same user (or privileged users
like root), so

    escevalidenv ESCEVALED_SECRET

is strictly more secure than

    escevalid "$ESCEVALED_SECRET"

The biggest subtle value is that when you intend
to access an environment variable, it's explicit
and no one has to know if the current shell can
and is shadowing the environment variable with a
non-exported shell variable of the same name (no
modern shell does this as far as I know, but it
ought to be entirely irrelevant/decoupled).

The biggest practical value is just convenience;
easier+faster to type, just letters and shift
instead of special characters.