Better practices in escevalid
This is closer to how I think control flow and
parsing/validating functions should generally
be written: don't exit the program leafwards
in the call tree, instead return an indication
of success or failure which can neatly compose
in the language's conditional statements.

Ideally parsers/validators also provide enough
information to precisely identify where the
parse/validation failed and why, but for a
project with so little usage, that's effort
and complexity for another commit, if ever.