Quote adjacent quotes instead of backslashing
Just implemented the C version to try this out.
Committing this so that it can be retrieved if
needed, but I currently don't intend to use it.

This encoding is strictly more efficient, but
the difference only arises when there are at
least three single-quote characters next to
each other ("''" substrings occur whenever an
escevall'ed string is esceval'ed again, but
"'''" don't - at least not now that redundant
quote pairs are eliminated). In the rare case
that this happens, the benefit is O(1) vs
O(number of adjacent single quotes). The
benefit for additional layers of esceval is
pretty decent too (instead of each quote
ballooning into multiple characters which also
all balloon on the next encoding, the original
quotes stay constant and only the wrapping
characters balloon), but in practice the depth
of esceval encoding is rarely high enough (and
if we're at scale where this difference
matters, it's probablu time to use something
else for encoding and handling data than
esceval and shell).