Revert "Quote adjacent quotes instead of backslashing"
Besides the value being relatively low and the
implementation needing to be more complex, the
other big downside after trying it out is that
it's more complex and hard to read in all the
cases where it starts to make a big difference
for efficiency. Suddenly there is a third
special character (`"`), the quotes tend to be
harder to visually pick out (`'"'"''"'"'` is a
common substring, for example), and we need
two "algorithms" in our thoughts to parse out
single-quotes (either `\` or `"` can introduce
an escaped `'`).

This reverts commit 2646def76345b44184dd00270684d15a8049c8fb.