Fix edge case in escevalenv.sh
Discovered by the new testing. Nice. In
principle I should've caught this. Years
ago I prided myself on always having these
kinds of edge-cases in my thoughts whenever
I was shell-scripting. But now? I'm not so
sure that's a good use of my human brain.

Little-known, rarely-encountered quirk of
Bourne-shell command substitution: trailing
newlines are stripped off the output of the
command. Any number of them, in most shells.

So if, for example, you have a value with a
newline at the end, and you, say, `printenv`
it (`printenv` writes the value followed by
one newline), both `printenv`'s newline and
the value's newline are stripped off when
capturing that output in a variable.

The robust/general solution is seen in this
commit: we echo another character inside
the command substitution after the desired
output, which shields all trailing newlines
in the `printenv` output, then delete that
character and one trailing newline (added
by `printenv`) afterwards.