Rename escevalid to escevalcheck
This portmanteau name was probably a little too
clever/artsy, especially when considering an
international audience, and the fact that it's
too ambiguous with other just-plausible-enough
parses+interpretations like "esceval ID". A name
shouldn't need an explanation every time it's
introduced to ensure correct interpretation and
pronounciation and so on.